+++ src/popup/components/Navbar/Sidebar/Sidebar.vue
@@ -1,13 +1,14 @@
 <script setup lang="ts">
 import { useSidebarStore } from '@/stores/sidebar'
 import { useTransStore } from '@/stores/trans'
+import openSettingsPage from '@/modules/openSettingsPage'
 import isFirefox from '@common/modules/isFirefox'
 import getImageURL from '@common/modules/getImageURL'
 import Hamburger from '@/components/Navbar/Sidebar/Hamburger.vue'
 import RightSlideTransition from '@common/components/Transitions/RightSlideTransition.vue'
 import Overlay from '@/components/Navbar/Sidebar/Overlay.vue'
-import UserCircleIcon from '@common/components/Icons/UserCircleIcon.vue'
 import StarIcon from '@common/components/Icons/StarIcon.vue'
+import CogIcon from '@common/components/Icons/CogIcon.vue'
 import SidebarLink from '@/components/Navbar/Sidebar/SidebarLink.vue'
 import Heading from '@/components/Navbar/Sidebar/Heading.vue'
 import GitHubIcon from '@common/components/Icons/GitHubIcon.vue'
@@ -20,11 +21,6 @@ const rateUsLink = isFirefox()
     : 'https://chromewebstore.google.com/detail/tab-guardian/kjdklkfpliphcbnphmfhalllclfieojp'
 
 const links = [
-    {
-        href: 'https://serhii.io/about-me',
-        label: trans('Extension Author'),
-        icon: UserCircleIcon,
-    },
     {
         href: 'https://github.com/tab-guardian/tab-guardian',
         label: trans('Contribute'),
@@ -39,6 +35,12 @@ const links = [
         href: rateUsLink,
         label: trans('Rate Us'),
         icon: StarIcon,
+    },
+    {
+        href: 'javascript:',
+        label: trans('Settings'),
+        icon: CogIcon,
+        callback: openSettingsPage,
     },
 ]
 </script>
@@ -66,6 +68,7 @@ const links = [
                         :href="link.href"
                         :label="link.label"
                         :icon="link.icon"
+                        @click="link.callback"
                     />
                 </div>