+++ src/popup/components/Views/GroupView/GroupName.vue
@@ -4,9 +4,7 @@ import type { Group } from '@/types'
 import { useGroupStore } from '@/stores/group'
 import GroupIcon from '@/components/Views/MainView/Groups/GroupIcon.vue'
 
-defineProps<{
-    group: Group
-}>()
+defineProps<{ group: Group }>()
 
 const classes = 'text-lg my-1 px-2 py-0.5'
 const inputRef = ref<HTMLInputElement | null>(null)
@@ -24,19 +22,28 @@ watchEffect(() => {
     <div class="flex items-center gap-1 relative">
         <GroupIcon v-if="group.icon" :group />
 
-        <input
+        <form
             v-if="store.isTitleFieldActive"
-            v-model="store.newGroup.name"
-            @blur="store.renameGroup"
-            :class="classes"
-            :maxlength="store.groupNameMaxLength"
-            ref="inputRef"
-            type="text"
-            class="border border-border bg-safe text-font w-full rounded-md pr-16"
-            autofocus
-        />
+            @submit.prevent="store.isTitleFieldActive = false"
+        >
+            <input
+                v-model="store.newGroup.name"
+                @blur="store.renameGroup"
+                :class="classes"
+                :maxlength="store.groupNameMaxLength"
+                ref="inputRef"
+                type="text"
+                class="border border-border bg-safe text-font w-full rounded-md pr-16"
+                autofocus
+            />
+        </form>
 
-        <h2 v-else :class="classes" class="border border-transparent">
+        <h2
+            v-else
+            v-on:dblclick="store.startGroupRenaming"
+            :class="classes"
+            class="border border-transparent select-none"
+        >
             {{ group.name }}
         </h2>