+++ src/settings/App.vue
@@ -4,17 +4,19 @@ import { useSettingsStore } from '@/stores/settings'
 import { useTransStore } from '@/stores/trans'
 import en from '@settings/locales/en.json'
 import ru from '@settings/locales/ru.json'
+import zh from '@settings/locales/zh.json'
 import Spinner from '@/components/Spinner.vue'
 import OptionsSection from '@settings/components/OptionsSection.vue'
 import EraseSection from '@settings/components/EraseSection.vue'
 import ExportSection from '@settings/components/ExportSection.vue'
 import ImportSection from '@settings/components/ImportSection.vue'
+import LocaleSection from '@settings/components/LocaleSection.vue'
 
 const store = useSettingsStore()
 const { trans, loadMessages } = useTransStore()
 
-onBeforeMount(() => {
-    loadMessages({ en, ru })
+onBeforeMount(async () => {
+    await loadMessages({ en, ru, zh })
 })
 
 onMounted(() => {
@@ -32,7 +34,10 @@ onMounted(() => {
         <Spinner v-if="store.loading" />
 
         <div v-else class="grid grid-cols-1 lg:grid-cols-2 gap-5 mt-4">
-            <OptionsSection />
+            <div class="flex flex-col gap-4">
+                <LocaleSection />
+                <OptionsSection />
+            </div>
 
             <div class="flex flex-col gap-4">
                 <EraseSection />